History - Add clear button

+++ sflphone-gtk/src/dbus.h
@@ -313,12 +313,12 @@ void dbus_start_hidden( void );
  * ConfigurationManager - Gives the maximum number of calls the user wants to have in the history 
  * @return double The maximum number of calls
  */
-gdouble dbus_get_max_calls( void );
+guint dbus_get_max_calls( void );
 
 /**
  * ConfigurationManager - Gives the maximum number of calls the user wants to have in the history 
  */
-void dbus_set_max_calls( const gdouble calls );
+void dbus_set_max_calls( const guint calls );
 /**
  * ConfigurationManager - Configure the start-up option
  * @return int	1 if SFLphone should start in the system tray
